+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **ThrustController** containing most of the settings from the Utilities window
 - `LaunchToRendezvous()` and `LaunchToTargetPlane()` methods to AscentAutopilot
 - `AutostagingOnce` property to StagingController
+### Removed
+- Setters of `AutoTurnStartAltitude`, `AutoTurnStartVelocity`, `AutoTurnEndAltitude` in **AscentClassic** - these
+properties should be read-only
 ### Fixed
 - `AscentAutopilot#DesiredInclination` property didn't do anything